Changing type of the max buffer size for roctracer to int64 instead of int32.

+++ libkineto/src/RoctracerActivityApi.h
@@ -45,7 +45,7 @@ class RoctracerActivityApi {
   void clearActivities();
   void flushActivities();
   void teardownContext() {}
-  void setMaxEvents(uint32_t maxEvents);
+  void setMaxEvents(uint64_t maxEvents);
 
   virtual int processActivities(
       std::function<void(const roctracerBase*)> handler,
